feat: Adicionar Middleware nas rotas

+++ src/Routes/Products/Cart/routes.ts
@@ -8,9 +8,10 @@ import { EditCartItemsController } from "../../../Controllers/Products/Cart/edit
 
 const cartRoutes: Router = Router()
 
-cartRoutes.route("/carts/register-cart-items").post((req: Request, res: Response) =>{RegisterCartController.register(req, res)})
-cartRoutes.route("/carts/get-cart-items/:id_user_fk").get((req: Request, res: Response) =>{GetCartDatasController.getCartDatas(req, res)})
-cartRoutes.route("/carts/delete-cart-items/:id_user_fk").delete((req: Request, res: Response) =>{DeleteCartItemController.deleteCartItem(req, res)})
-cartRoutes.route("/carts/delete-all-cart-items/:id_cart").delete((req: Request, res: Response) => {DeleteAllCarItemsController.deleteAllCartItems(req, res)})
-cartRoutes.route("/carts/edit-cart-items/:id_user_fk").put((req: Request, res: Response) =>{EditCartItemsController.editCartItems(req, res)})
+cartRoutes.route("/carts/register-cart-items").post(MiddlewareAuthorization.authorization, MiddlewareAuthorization.isClient, (req: Request, res: Response) =>{RegisterCartController.register(req, res)})
+cartRoutes.route("/carts/get-cart-items/:id_user_fk").get(MiddlewareAuthorization.authorization, MiddlewareAuthorization.isClient, (req: Request, res: Response) =>{GetCartDatasController.getCartDatas(req, res)})
+cartRoutes.route("/carts/delete-cart-items/:id_user_fk").delete(MiddlewareAuthorization.authorization, MiddlewareAuthorization.isClient, (req: Request, res: Response) =>{DeleteCartItemController.deleteCartItem(req, res)})
+cartRoutes.route("/carts/delete-all-cart-items/:id_cart").delete(MiddlewareAuthorization.authorization, MiddlewareAuthorization.isClient, (req: Request, res: Response) => {DeleteAllCarItemsController.deleteAllCartItems(req, res)})
+cartRoutes.route("/carts/edit-cart-items/:id_user_fk").put(MiddlewareAuthorization.authorization, MiddlewareAuthorization.isClient, (req: Request, res: Response) =>{EditCartItemsController.editCartItems(req, res)})
+
 export {cartRoutes}
